% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity.R
\name{complexity}
\alias{complexity}
\alias{complexity.default}
\alias{complexity.formula}
\title{Extract the complexity measures from datasets}
\usage{
complexity(...)

\method{complexity}{default}(x, y, groups = "all", ...)

\method{complexity}{formula}(formula, data, groups = "all", ...)
}
\arguments{
\item{...}{Not used.}

\item{x}{A data.frame contained only the input attributes.}

\item{y}{A response vector with one value for each row/component of x.}

\item{groups}{A list of complexity measures groups or \code{"all"} to include
all of them.}

\item{formula}{A formula to define the output column.}

\item{data}{A data.frame dataset contained the input and output attributes.}

\item{type}{The type of supervised problem: The \code{"class"} is used for 
classification tasks and \code{"regr"} for regression tasks.}
}
\value{
A numeric vector named by the requested complexity measures.
}
\description{
This function is responsable to extract the complexity measures from the 
classes of the classification tasks. For such, they take into account the 
overlap between classes imposed by feature values, the separability and 
distribution of the data points.
}
\details{
The following groups are allowed for this method:
 \describe{
   \item{"overlapping"}{The feature overlapping measures characterize how 
     informative the available features are to separate the classes See 
     \link{overlapping} for more details.}
   \item{"neighborhood"}{Neighborhood measures characterize the presence and 
     density of same or different classes in local neighborhoods. See 
     \link{neighborhood} for more details.}
   \item{"linearity"}{Linearity measures try to quantify whether the classes 
     can be linearly separated. See \link{linearity.class} or 
     \link{linearity.regr} for more details.}
 }
}
\examples{
## Extract all complexity measures for classification task
data(iris)
complexity(Species ~ ., iris, type="class")
}
\references{
Victor H. Barella and Luis P. Garcia and Marcilio P. de Souto and Ana C. Lorena
    and Andre de Carvalho (2018). Data Complexity Measures for Imbalanced
    Classification Tasks. In 2018 International Joint Conference on Neural Networks
    (IJCNN) (pp. 1-8). IEEE.

 Tin K Ho and Mitra Basu. (2002). Complexity measures of supervised 
   classification problems. IEEE Transactions on Pattern Analysis and Machine
   Intelligence, 24, 3, 289--300.

 Albert Orriols-Puig, Nuria Macia and Tin K Ho. (2010). Documentation for 
   the data complexity library in C++. Technical Report. La Salle - 
   Universitat Ramon Llull.

 Ana C Lorena and Aron I Maciel and Pericles B C Miranda and Ivan G Costa and
   Ricardo B C Prudencio. (2018). Data complexity meta-features for 
   regression problems. Machine Learning, 107, 1, 209--246.
}
