% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImbCoL-package.R
\docType{package}
\name{ImbCoL-package}
\alias{ImbCoL}
\alias{ImbCoL-package}
\title{ImbCoL: Data Complexity Measures for Imbalanced Classification Tasks}
\description{
These are a set of measures to assess the difficulty of each class of a classification task based on the training set. The data must be structured in a data.frame format. For more information about the data complexity measures decomposed by class, check Barella et al. (doi: 10.1109/IJCNN.2018.8489661)
}
\author{
\strong{Maintainer}: Victor Barella \email{victorhb@icmc.usp.br}

}
\keyword{internal}
